--- __mocks__/PrivacyPolicy.js
@@ -1,7 +1,0 @@
-export default `
-
-# Privacy Policy
-
-This is a mock implementation of the privacy policy file for unit testing.
-
-`;
