+++ src/__tests__/src/PrivacyPolicy.test.tsx
@@ -12,13 +12,22 @@ jest.mock('react-router-dom', () => ({
 window.scrollTo = jest.fn();
 
 describe('Privacy Policy', () => {
+  it('it has the title', () => {
+    const { getByRole } = render(<PrivacyPolicy />);
+    expect(getByRole('button').textContent).toContain('Go Back');
+  });
+
   it('it has the back button', () => {
-    const { queryAllByText, getByRole } = render(<PrivacyPolicy />);
+    const { getByRole } = render(<PrivacyPolicy />);
     expect(getByRole('button').textContent).toContain('Go Back');
   });
 
   it('it has the privacy policy text', () => {
-    const { getByTestId } = render(<PrivacyPolicy />);
-    expect(getByTestId('PrivacyPolicyText'));
+    const { getByText } = render(<PrivacyPolicy />);
+    expect(
+      getByText(
+        /All we know about you is what you share by visiting and exploring our website. We don't collect any of your personal information from any third parties/i
+      )
+    );
   });
 });
